fixed duplicate term issue in search history

+++ src/main/frontend/search-history.js
@@ -2,11 +2,22 @@
 function _updateHistoryFromHashChange(event) {
   var recent = JSON.parse($.localStorage.getItem('search_history_recent_list')) || [];
   if (event.getState( 'query' )) {
-    var new_len = recent.unshift(event.getState( 'query' ));
-    if (new_len > 10) {
-      recent.pop();
-    };
-    $.localStorage.setItem('search_history_recent_list', JSON.stringify(recent));
+    var new_term = event.getState( 'query' );
+	var already_exist = false;
+	$.each(recent, function(index, value){
+	  if(value == new_term){
+		already_exist = true;
+		return false; // break the loop
+	  }
+    });
+	if(!already_exist){
+	  var new_len = recent.length;
+	  new_len = recent.unshift(new_term);
+      if (new_len > 10) {
+        recent.pop();
+      }
+      $.localStorage.setItem('search_history_recent_list', JSON.stringify(recent));
+	}
   };
   $( document ).ready( function() {
     $ ( "#historyTable .dynamicrow" ).remove();
